+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/SdkSynchronizer.kt
@@ -17,6 +17,7 @@ import cash.z.ecc.android.sdk.exception.TransactionEncoderException
 import cash.z.ecc.android.sdk.exception.TransactionSubmitException
 import cash.z.ecc.android.sdk.ext.ConsensusBranchId
 import cash.z.ecc.android.sdk.ext.ZcashSdk
+import cash.z.ecc.android.sdk.internal.FastestServerFetcher
 import cash.z.ecc.android.sdk.internal.SaplingParamTool
 import cash.z.ecc.android.sdk.internal.Twig
 import cash.z.ecc.android.sdk.internal.TypesafeBackend
@@ -25,6 +26,7 @@ import cash.z.ecc.android.sdk.internal.block.CompactBlockDownloader
 import cash.z.ecc.android.sdk.internal.db.DatabaseCoordinator
 import cash.z.ecc.android.sdk.internal.db.derived.DbDerivedDataRepository
 import cash.z.ecc.android.sdk.internal.db.derived.DerivedDataDb
+import cash.z.ecc.android.sdk.internal.exchange.UsdExchangeRateFetcher
 import cash.z.ecc.android.sdk.internal.ext.existsSuspend
 import cash.z.ecc.android.sdk.internal.ext.isNullOrEmpty
 import cash.z.ecc.android.sdk.internal.ext.tryNull
@@ -41,6 +43,9 @@ import cash.z.ecc.android.sdk.internal.transaction.TransactionEncoder
 import cash.z.ecc.android.sdk.internal.transaction.TransactionEncoderImpl
 import cash.z.ecc.android.sdk.model.Account
 import cash.z.ecc.android.sdk.model.BlockHeight
+import cash.z.ecc.android.sdk.model.FastestServersResult
+import cash.z.ecc.android.sdk.model.FetchFiatCurrencyResult
+import cash.z.ecc.android.sdk.model.ObserveFiatCurrencyResult
 import cash.z.ecc.android.sdk.model.PercentDecimal
 import cash.z.ecc.android.sdk.model.Proposal
 import cash.z.ecc.android.sdk.model.TransactionOverview
@@ -63,20 +68,28 @@ import co.electriccoin.lightwallet.client.new
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.flow.channelFlow
 import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
@@ -100,13 +113,15 @@ import kotlin.time.Duration.Companion.seconds
  * @property processor saves the downloaded compact blocks to the cache and then scans those blocks for
  * data related to this wallet.
  */
-@Suppress("TooManyFunctions")
+@Suppress("TooManyFunctions", "LongParameterList")
 class SdkSynchronizer private constructor(
     private val synchronizerKey: SynchronizerKey,
     private val storage: DerivedDataRepository,
     private val txManager: OutboundTransactionManager,
     val processor: CompactBlockProcessor,
-    private val backend: TypesafeBackend
+    private val backend: TypesafeBackend,
+    private val fetchFastestServers: FastestServerFetcher,
+    private val fetchExchangeChangeUsd: UsdExchangeRateFetcher
 ) : CloseableSynchronizer {
     companion object {
         private sealed class InstanceState {
@@ -136,7 +151,9 @@ class SdkSynchronizer private constructor(
             repository: DerivedDataRepository,
             txManager: OutboundTransactionManager,
             processor: CompactBlockProcessor,
-            backend: TypesafeBackend
+            backend: TypesafeBackend,
+            fastestServerFetcher: FastestServerFetcher,
+            fetchExchangeChangeUsd: UsdExchangeRateFetcher,
         ): CloseableSynchronizer {
             val synchronizerKey = SynchronizerKey(zcashNetwork, alias)
 
@@ -149,7 +166,9 @@ class SdkSynchronizer private constructor(
                     repository,
                     txManager,
                     processor,
-                    backend
+                    backend,
+                    fastestServerFetcher,
+                    fetchExchangeChangeUsd
                 ).apply {
                     instances[synchronizerKey] = InstanceState.Active
                     start()
@@ -188,13 +207,50 @@ class SdkSynchronizer private constructor(
         }
     }
 
-    private val _status = MutableStateFlow<Synchronizer.Status>(DISCONNECTED)
+    private val _status = MutableStateFlow(DISCONNECTED)
 
-    var coroutineScope: CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main)
+    val coroutineScope: CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main)
 
     override val orchardBalances = processor.orchardBalances.asStateFlow()
     override val saplingBalances = processor.saplingBalances.asStateFlow()
     override val transparentBalance = processor.transparentBalance.asStateFlow()
+
+    private val refreshExchangeRateUsd = MutableSharedFlow<Unit>(replay = 1).apply { tryEmit(Unit) }
+
+    private var lastExchangeRateValue = ObserveFiatCurrencyResult()
+
+    @OptIn(ExperimentalCoroutinesApi::class)
+    override val exchangeRateUsd =
+        channelFlow {
+            refreshExchangeRateUsd
+                .flatMapLatest {
+                    flow {
+                        emit(lastExchangeRateValue.copy(isLoading = true))
+                        lastExchangeRateValue =
+                            when (val result = fetchExchangeChangeUsd()) {
+                                is FetchFiatCurrencyResult.Error -> lastExchangeRateValue.copy(isLoading = false)
+
+                                is FetchFiatCurrencyResult.Success ->
+                                    lastExchangeRateValue.copy(
+                                        isLoading = false,
+                                        currencyConversion = result.currencyConversion
+                                    )
+                            }
+                        emit(lastExchangeRateValue)
+                    }
+                }
+                .onEach { send(it) }
+                .flowOn(Dispatchers.Default)
+                .launchIn(this)
+
+            awaitClose {
+                // do nothing
+            }
+        }.flowOn(Dispatchers.Default).stateIn(
+            scope = coroutineScope,
+            started = SharingStarted.WhileSubscribed(), // stop immediately
+            initialValue = ObserveFiatCurrencyResult()
+        )
 
     override val transactions
         get() =
@@ -296,6 +352,11 @@ class SdkSynchronizer private constructor(
 
     override val latestBirthdayHeight
         get() = processor.birthdayHeight
+
+    override suspend fun getFastestServers(
+        context: Context,
+        servers: List<LightWalletEndpoint>
+    ): Flow<FastestServersResult> = fetchFastestServers(context, servers)
 
     internal fun start() {
         coroutineScope.onReady()
@@ -415,6 +476,12 @@ class SdkSynchronizer private constructor(
      */
     suspend fun refreshAllBalances() {
         processor.refreshWalletSummary()
+    }
+
+    override suspend fun refreshExchangeRateUsd() {
+        coroutineScope.launch {
+            refreshExchangeRateUsd.emit(Unit)
+        }
     }
 
     suspend fun isValidAddress(address: String): Boolean {
@@ -542,7 +609,7 @@ class SdkSynchronizer private constructor(
         val reason = if (scannedRange.isNullOrEmpty()) "it's been a while" else "new blocks were scanned"
 
         if (shouldRefresh) {
-            Twig.info { "Triggering utxo refresh since $reason!" }
+            Twig.debug { "Triggering utxo refresh since $reason!" }
             refreshUtxos(Account.DEFAULT)
 
             Twig.debug { "Triggering balance refresh since $reason!" }
@@ -636,6 +703,7 @@ class SdkSynchronizer private constructor(
                         is TransactionSubmitResult.Success -> {
                             // Expected state
                         }
+
                         is TransactionSubmitResult.Failure,
                         is TransactionSubmitResult.NotAttempted -> {
                             anySubmissionFailed = true
@@ -673,6 +741,7 @@ class SdkSynchronizer private constructor(
             is TransactionSubmitResult.Success -> {
                 return storage.findMatchingTransactionId(encodedTx.txId.byteArray)!!
             }
+
             else -> {
                 throw TransactionSubmitException()
             }
@@ -708,6 +777,7 @@ class SdkSynchronizer private constructor(
             is TransactionSubmitResult.Success -> {
                 return storage.findMatchingTransactionId(encodedTx.txId.byteArray)!!
             }
+
             else -> {
                 throw TransactionSubmitException()
             }
@@ -764,6 +834,7 @@ class SdkSynchronizer private constructor(
                     Twig.info { "Chain tip for validate consensus branch action fetched: ${response.result.value}" }
                     runCatching { response.result.toBlockHeight(network) }.getOrNull()
                 }
+
                 is Response.Failure -> {
                     Twig.error {
                         "Chain tip fetch failed for validate consensus branch action with:" +
@@ -841,6 +912,7 @@ class SdkSynchronizer private constructor(
                         return ServerValidation.InValid(it)
                     }
                 }
+
                 is Response.Failure -> {
                     return ServerValidation.InValid(response.toThrowable())
                 }